9. Vervaardig RecipeInfo class Fixes #8

+++ lib/gerecht_info.php
@@ -1,52 +1,70 @@
-<?php 
+<?php
+
 class gerecht_info{
-
     private $connection;
+    private $user;
     public function __construct($connection){
-        $this->connection =$connection;
+        $this->connection=$connection;
+        $this->user=new user($connection);
     }
-    public function selecteerInfo($gerecht_info_id){
-        $sql="select * from gerecht_info 
-                LEFT JOIN user
-                ON user.id = gerecht_info.user_id
-                WHERE gerecht_info.gerecht_id = $gerecht_info_id";
-                
 
-        $result=mysqli_query($this->connection,$sql);
-        $gerecht_info=mysqli_fetch_array($result,MYSQLI_ASSOC);
+public function selecteerInfo($gerecht_id){
+    $sql="select * from gerecht_info where gerecht_id=$gerecht_id";
 
-        return($gerecht_info);
-    }
-   public function addFavorite($gerecht_id, $user_id) {
+    $return=[];
 
-    $sql = "
-        INSERT INTO gerecht_info (record_type_id, gerecht_id, user_id)
-        VALUES ('F', $gerecht_id, $user_id)
-        
-    ";
-
-    return mysqli_query($this->connection, $sql);
+    $result=mysqli_query($this->connection,$sql);
     
+    while ($row=mysqli_fetch_array($result,MYSQLI_ASSOC)){
+        $user=$this->selecteerUser($row["user_id"]);
 
-    if ($obj->addFavorite($gerecht_id, $user_id)) {
-    echo "Recipe $gerecht_id is favorited by $user_id";
-} else {
-    echo "Error: " . mysqli_error($connection);
-}
+        if($row["record_type"]=="O"||$row["record_type"]=="F" ){
+
+            $return[]= [
+            "id"=>$row["id"],
+            "record_type"=>($row["record_type"]),
+            "gerecht_id"=>($row["gerecht_id"]),
+            "user_id"=>($user["id"]),
+            "naam"=>($user["naam"]),
+            "email"=>($user["email"]),
+            "wachtwoord"=>($user["wachtwoord"]),
+            "afbeelding"=>($user["afbeelding"]),
+            "datum"=>($row["datum"]),
+            "nummeriekveld"=>($row["nummeriekveld"]),
+            "tekstveld"=>($row["tekstveld"]),
+            
+        ];
+           
+        }else{
+        $return[]= $row;
+        }
     }
-
-    public function deleteFavorite($gerecht_id, $user_id) {
-
-    $sql = "
-        DELETE FROM gerecht_info
-        WHERE record_type_id = 'F'
-          AND gerecht_id = $gerecht_id
-          AND user_id = $user_id
-    ";
-
-    return mysqli_query($this->connection, $sql);
-    
-    }
-
+    return $return;
 }
+
+private function selecteerUser($user_id){
+
+    return $this->user->selecteerUser($user_id);
+}
+
+private function addFavorite($user_id,$gerecht_id){
+    $sql= "INSERT INTO gerecht_info (user_id, gerecht_id, record_type) VALUES
+    ('$user_id','$gerecht_id','F')"; 
+
+    $result =mysqli_query($this->connection,$sql);
+    $favorited=mysqli_fetch_array($result,MYSQLI_ASSOC);
+
+    return($favorited);
+}
+
+private function deleteFavorite($user_id,$gerecht_id){
+     $sql= "DELETE FROM gerecht_info WHERE user_id=$user_id AND gerecht_id=$gerecht_id AND record_type='F'";
+
+    $result =mysqli_query($this->connection,$sql);
+    $unfavorited=mysqli_fetch_array($result,MYSQLI_ASSOC);
+
+    return($unfavorited);
+}
+}
+
 ?>
